refactor(curriculum): co-locate WorkProgram role predicates in role_predicates.go
Closes Tier 2 carry-forward from v0.178.0 code-review (SHIP 9.50/10
— Cohesion axis flagged "four predicates spread across four files,
which is a cohesion smell" at the 6+ predicate horizon).

Extract the four role-matrix predicates into a dedicated file:
- isAllowedToCreateWorkProgram (was in create_work_program_usecase.go)
- isAuthorOrSystemAdmin         (was in submit_work_program_usecase.go)
- isApprover                    (was in approve_work_program_usecase.go)
- canViewWorkProgram            (was in get_work_program_usecase.go)

The four predicates together form the ADR-018 ADR-5 role decision
table, so reading them side-by-side makes the access-control matrix
auditable as a single artefact rather than as a grep across four
files. The package docstring at the top of role_predicates.go anchors
each predicate to its use-case consumer for navigation.

The original use-case files lose only their imports of authDomain
(no longer needed there) and the predicate bodies. Behaviour is
unchanged; tests pass.
